Docker Cloud/Docker Multi-tenant/api-gateway/requirements.txt
flask==2.3.3
requests==2.31.0
PyJWT==2.8.0